Estimate Difference-in-Difference Causal Forests for Long Format Data

This function computes the difference outcome $$Z_{it} = Y_{it} - Y_{ib}$$ for each time period \( t \) (using base period \( b \)) and estimates a causal forest using the grf package on \((Z, W, X)\). Y is assumed to be in long format. Therefore, you must provide the corresponding time and id vectors. It is required that every unit (as identified by id) has an observation in the base period \( b \).

Usage

estimate_DiDCF(Y, W, X, b, time, id, ...)

Arguments

Y

A numeric vector of the outcome variable.

W

A vector of treatment indicators (one per unit).

X

A numeric matrix or data frame of features that vary only in the cross-sectional dimension (one row per unit).

b

The value in the time vector corresponding to the base period.

time

A vector indicating the time period for each observation in Y.

id

A vector indicating the unit identifier for each observation in Y.

...

Additional arguments to be forwarded to grf::causal_forest.

Value

A list of causal forest objects (one for each time period other than the base period).

Details

Additional arguments (...) are passed directly to the call to grf::causal_forest.

Examples

estimate_DiDCF(example_data$Y,
               example_data$t_indicator[example_data$period==1],
               as.data.frame(model.matrix(~.,data=example_data[example_data$period==1,c("x_1","x_2")])),
               1,
               data$period,
               data$unit_id)
#> Error in estimate_DiDCF(example_data$Y, example_data$t_indicator[example_data$period ==     1], as.data.frame(model.matrix(~., data = example_data[example_data$period ==     1, c("x_1", "x_2")])), 1, data$period, data$unit_id): could not find function "estimate_DiDCF"